requirements.txt
pycryptodome>=3.18.0
flask>=2.0.1
flask-cors>=3.0.10 